test: 멤버 등급 조회 테스트 코드 작성
